scraper/requirements.txt
# Requirements pour le module scraper
yt-dlp>=2024.3.10
requests>=2.28.2
python-dotenv>=1.0.0
tqdm>=4.65.0
pymongo>=4.3.3
